+++ src/controllers/ProjectController.ts
@@ -12,6 +12,8 @@ import ProjectService from "@/services/ProjectService";
 
 import BaseController from "./BaseController";
 
+export type ProjectInputSchema = Omit<Project, keyof HiddenBodyKeys>;
+
 @Tags("Project")
 @Route("project")
 export default class ProjectController extends BaseController<Project> {
@@ -19,24 +21,28 @@ export default class ProjectController extends BaseController<Project> {
 		super(new ProjectService());
 	}
 
+	@Security("api_key")
 	@Security("jwt")
 	@Get("/")
 	read(@Queries() queryParams?: IGetQueryParams) {
 		return super.read();
 	}
 
+	@Security("api_key")
 	@Security("jwt")
 	@Post("/")
-	create(@Body() body: Omit<Project, keyof HiddenBodyKeys>, @Queries() queryParams?: IPostQueryParams) {
+	create(@Body() body: ProjectInputSchema, @Queries() queryParams?: IPostQueryParams) {
 		return super.create(body);
 	}
 
+	@Security("api_key")
 	@Security("jwt")
 	@Patch("/")
-	update(@Body() body: Omit<Project, keyof HiddenBodyKeys>, @Queries() queryParams?: IPostQueryParams) {
+	update(@Body() body: ProjectInputSchema, @Queries() queryParams?: IPostQueryParams) {
 		return super.update(body);
 	}
 
+	@Security("api_key")
 	@Security("jwt")
 	@Delete("/")
 	async delete(@Queries() queryParams?: IDeleteQueryParams) {
@@ -71,6 +77,7 @@ export default class ProjectController extends BaseController<Project> {
 		return super.delete();
 	}
 
+	@Security("api_key")
 	@Security("jwt")
 	@Get("/with-apps")
 	async getProjectsAndApps(@Queries() queryParams?: IGetQueryParams) {