+++ src/plugins/os.ts
@@ -1,6 +1,5 @@
-var osvar = process.platform;
-
 export function getOS() {
+	const osvar = process.platform;
 	if (osvar == "darwin") {
 		return "mac";
 	} else if (osvar == "win32") {
@@ -13,13 +12,16 @@ export function getOS() {
 }
 
 export function isWin() {
+	const osvar = process.platform;
 	return osvar == "win32";
 }
 
 export function isMac() {
+	const osvar = process.platform;
 	return osvar == "darwin";
 }
 
 export function isLinux() {
+	const osvar = process.platform;
 	return osvar == "darwin";
 }